MNT: Moving some imports as well as added qvp imports.
