+++ pyart/retrieve/__init__.py
@@ -17,6 +17,7 @@ from .qpe import est_rain_rate_zpoly, est_rain_rate_z, est_rain_rate_kdp
 from .qpe import est_rain_rate_a, est_rain_rate_zkdp, est_rain_rate_za
 from .qpe import est_rain_rate_hydro
 from .vad import vad_michelson, vad_browning
-from .qvp import quasi_vertical_profile
+from .qvp import quasi_vertical_profile, compute_qvp_ms, compute_rqvp
+from .qvp import compute_evp, compute_svp, compute_vp
 
 __all__ = [s for s in dir() if not s.startswith('_')]
